refactor(dashboard-tsr): split agent-visualization into folder (#1392)
Extract types/helpers, chart views, and chart controls into a co-located
agent-visualization/ folder; dedup the duplicate series-config memo and
X-axis blocks into shared helpers. Public AgentVisualization +
AgentVisualizationProps/VizConfig exports unchanged.

Part of epic #1392
